+++ src/style/select.less
@@ -739,7 +739,13 @@ details {
   .label, .mut-count {
     opacity: 1;
     transition: opacity 0.15s;
+    font-size: 12;
 
+  .label-bg {
+    fill: white;
+    fill-opacity: 0.7;
+    }
+    
     // .name {
     //   stroke: none;
     // }
