+++ src/views/select.html
@@ -77,9 +77,10 @@
               <svg width="256" height="160" viewBox="0 0 256 160">
                 <path class="connector" d="M0 0 L-45 0 -45 -15"></path>
                 <g class="label">
-                  <ellipse class="outline" cx="-15" cy="-14" rx="30" ry="28"></ellipse>
-                  <ellipse class="name" cx="-11" cy="-10" rx="22" ry="20"></ellipse>
+                  <ellipse class="outline" cx="-15" cy="-15" rx="30" ry="30"></ellipse>
+                  <ellipse class="name" cx="-10" cy="-10" rx="20" ry="20"></ellipse>
                   <text x="0" y="0">A</text>
+                  <rect class="label-bg" x="0" y="0"></rect>
                 </g>
                 <g class="mut-count">
                   <rect x="-56" y="-10" width="22" height="20"></rect>
